# services/evaluation/internal/handler/docs/swagger.yaml
definitions:
...
  db.Project:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        $ref: '#/definitions/db.ProjectStatus'
    type: object
...
  db.ProjectStatus:
    enum:
    - ready
    - processing_remarks
    - processing_checklist
    - generating_final_report
    type: string
    x-enum-varnames:
    - ProjectStatusReady
    - ProjectStatusProcessingRemarks
    - ProjectStatusProcessingChecklist
    - ProjectStatusGeneratingFinalReport
  handler.Error:
    properties:
      error: {}
    type: object
...
  models.CreateProjectRequest:
    properties:
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
info:
  contact: {}
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Get list of all projects
      operationId: listProjects
      produces:
      - application/json
      responses:
        "200":
          description: List of projects
          schema:
            items:
              $ref: '#/definitions/db.Project'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.Error'
      summary: Get all projects
    post:
      consumes:
      - application/json
      description: Create a new project with name
      operationId: createProject
      parameters:
      - description: Project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            $ref: '#/definitions/db.Project'
        "400":
          description: Bad request - invalid input data
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.Error'
      summary: Create new project
...
swagger: "2.0"